# lib/api-zod/src/generated/types/createPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePostBodyStatus } from "./createPostBodyStatus";

export interface CreatePostBody {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  coverImage?: string | null;
  status: CreatePostBodyStatus;
  readTime: number;
}