# lib/api-zod/src/generated/types/listPostsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export type ListPostsResponse = Post[];