# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostStatus } from "./postStatus";

export interface Post {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  coverImage?: string | null;
  status: PostStatus;
  readTime: number;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}