# lib/api-zod/src/generated/types/updatePostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePostBodyStatus } from "./updatePostBodyStatus";

export interface UpdatePostBody {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  coverImage?: string | null;
  status?: UpdatePostBodyStatus;
  readTime?: number;
}